Add class BooksCollection, move book methods to Book and BooksCollection classes
