+++ book.rb
@@ -1,16 +1,18 @@
-require './rental'
-
 class Book
-  attr_reader :rentals
   attr_accessor :title, :author
 
   def initialize(title, author)
     @title = title
     @author = author
-    @rentals = []
   end
 
-  def add_rental(person, date)
-    Rental.new(date, self, person)
+  def self.create_book
+    print 'Title: '
+    title = gets.chomp
+    print 'Author: '
+    author = gets.chomp
+    book = Book.new(title, author)
+    puts 'Book created successfully!'
+    book
   end
 end
